vpn-provider: Support SplitRouting option from connmand
Support SplitRouting option in vpnd. The value is passed as
boolean and a warning is printed with invalid error if attempted to
be set as string. For setting the boolean values in provider settings
vpn_provider_set_boolean() is added.

Implement provider_property_changed() to be registered with every call
to connect_provider() for the signals sent about property changes on
VPN services. React only to "SplitRouting" value changes coming from
the VPN service via connmand.

Also always notify all components listening for the changes over VPN
D-Bus API on SplitRouting changes coming from connmand. The signal will
be reacted upon also by connmand but provider.c will ignore setting of
the same value to avoid D-Bus signal loop. For this the
vpn_provider_add_boolean() is modified to have an option to have th
signal sending forced regardless of the value change. This is required
for notifying the components listening for the changes via VPN D-Bus API
when connmand/connman-vpnd crashes and the value is not written yet to
the VPN service settings file and the component would not get an update,
for example, to "true" value it has if connmand and vpnd have false (by
default, vpnd has the value as false). There are duplicate signals sent
but this approach ensures that all components will know the correct
status of SplitRouting.

Do some minor cleanups to code in order to support code syntax.