[vpn-settings] Use definite user type in system user check. JB#50150
Rename vpn_settings_is_system_user() to vpn_settings_get_user_type() in
order to return the real result of the requested user. Add enum
vpn_settings_user_type for the values.

If the user is not found return VPN_SETTINGS_USER_TYPE_NOT_FOUND. Using
this in vpn-settings.c and vpnc.c to separate the system user from
normal user and unknown user fixes the errorous use in case the override
user is being set but the user in the config file for the VPN is not
found. If there is no override set for an invalid user simply return
NULL, otherwise use the override.

The VPN_SETTINGS_USER_TYPE_NOT_SET is for cases when there is no user
set in configs. In such case no override must be used if set as this
defaults to using root user.

Like before, normal user VPN_SETTINGS_USER_TYPE_REGULAR can be overridden
and system user VPN_SETTINGS_USER_TYPE_SYSTEM cannot.