[connman-vpn] Remove PrivateUsers option from service file. Contributes to JB#52008
Remove PrivateUsers option from connman-vpn.service because this breaks
connman-vpnd process access to the user's WLAN and VPN settings files.

When enabled this option creates new user namespace, but the process loses all
of the capabilities in the host (root) namespace. This means that connman-vpnd
loses CAP_DAC_IGNORE which prevents it from entering e.g. /home/defaultuser
directory. And because of the it can't access the saved WLAN and VPN settings,
so the connman service "forgets" all saved networks.

Also this option is not working on the old kernels like 3.10 making service
start fail.

Signed-off-by: Igor Zhbanov <i.zhbanov@omprussia.ru>