[service] Remove VPN transport dependency and improve sorting. JB#48769
There is no need to use the VPN transport dependency in service.c
anymore as plugins/vpn.c handle the transport checks. Also dnsproxy.c
manages the DNS servers using the split routing property.

Improved sorting to take account the lack of the transport dependency
and to use order when doing the sorting with preferred lists.