provider: Reformat IPv6 toggle function to use service.c functionality
[provider] Reformat IPv6 toggle function to use service.c funct. JB#53542

Moved the IPv6 toggling functionality for the most part to service.c and
this change adapts to the change. Only the VPN service checks are kept
in provider.c and IPv6 is disabled for all connected services instead of
using transport only and disabling IPv6 on system level. Current VPN and
transport services are passed to the service.c functionality to exclude
VPN service and always include transport service in IPv6 changes.

Changed to use "enable" to be in line with rest of the code. This
removes the need to use a lot of negations and perhaps it is easier to
understand the code as well.

Allow to change IPv6 only for VPNs that are connected (ready, as VPNs
cannot reach online state).

Prevent running the function as a loop within loop. This might happen
when current transport service is disconnected that in turn causes the
VPN to be disconnected as well. Then while looping through the services
another disconnect on the transport will get provider_indicate_state()
called again.

When re-enabling IPv6 enable the internal IPv6 suppor prior to resuming
IPv6 functionality on service basis. When disabling IPv6 do the internal
disabling after the services are handled. This is due to the changes
that were required for network.c to get new connections enabled but IPv6
disabled if IPv6 is internally disabled.