ipconfig: IPv6 method restore, force disabled to own func, always set autoconf
[ipconfig] IPv6 method restore, force disbled to own func, always set autoconf. JB#53542

Add functions to save and restore old IPv6 method. This is useful when
IPv6 is disabled termporarily for the time IPv4 VPN is used. First save
the IPv6 method before disabling and call restore when IPv6 needs to be
set up again using the old method.

Also, do not change system wide IPv6 disable_ipv6 with
__connman_ipconfig_set_ipv6_support(). Add also getter for the internal
IPv6 state.

Move the force disabled use from the enable/disable functions to own
separate getter and setter.

It is more straightforward to always disable/enable autoconf for IPv6
interface instead of managing it internally as well. This is a key
feature in disabling IPv6 along the disable_ipv6 value in /proc so
autoconf must be treated similarly as well. Dot save the value to be
more consistent in use.

When saving ipconfig use the original IPv6 method if the ipconfig is
force disabled. This ensures that "OFF" method is not saved to settings
file if the service and its ipconfig gets saved.

Remove accept_ra completely as unnecessary. autoconf is only required.