service: Support IPv6 enable/disable for connected services
[service] Support IPv6 enable/disable for conn services. JB#53542

Implement support for enabling/disabling IPv6 for every connected
service exluding the VPN and including the VPN transport. This is used
when an IPv4 VPN is connected to avoid leaking of data to IPv6 network
by disabling IPv6 for all services when connected and when disconnecting
the VPN re-enable IPv6.

When disabling the old IPv6 method is recorded, address is cleared, IPv6
network is disconnected, then turned to idle and notified. When enabling
IPv6 method is restored and simply enabling the ipconfig is required in
order to get the IPv6 connectivity to be resumed. In case the transport
has been disconnected prior to VPN disconnect it is imperative to do the
changes only to ipconfig to avoid changing the state of the transport.

Remove unnecessary __connman_service_notify_ipv6_configuration added in
5b7e6fc00e2a868c4f7bd3a7d45ca57b0b9d35bd

Change function name and use bool value as enable - remove negation.

Adapt to network.c changes in __connman_network_enable_ipconfig() to not
to include the force value. Instead use directly
__connman_ipconfig_set_force_disabled_ipv6() to set the ipconfig
internal value.