network: Prevent enabling IPv6 for network if internally disabled
[network] Dont' enable IPv6 for network if internally disabled. JB#53542

If a new connection comes up/is manually connected when IPv6 is
internally disabled for a IPv4 VPN the IPv6 configuration for the
network must be prevented. If the different IPv6 method functions do
this then an error is received from ipconfig.c which will propagate to
disabling of the new connection completely as network error is
triggered. This way IPv6 is silently ignored and forcefully set as
disabled to be re-enabled when VPN is disconnected.

Disable force disabled in IPv6 ipconfig when clearing the IPv6 address
in set_disconnect(). This is required when re-enabling IPv6 for kernel
to clear the values in order to return the proper /proc values for the
interface in case IPv6 was disabled for IPv4 VPN. Also restore the
original IPv6 method before setting disconnected to make kernel know
about the changes as well. Otherwise IPv6 on some occasions is not
enabled on the interface when, e.g., changing to another service of same
type technology.

Adopt to ipconfig.c changes to remove the force value from the IPv6
enable/disable functions. Instead use the direct ipconfig.c function to
set the value for force disabled. Thus, reverting most of the changes
done in 2031f6654e362cdb7bac2bb27979ecf5b36564ad