[provider] Allow to control VPN data leak prevention on provider basis. JB#53542
Add a function connman_provider_set_ipv6_data_leak_prevention() to be
used to control whether IPv6 is to be disabled on a provider when
connected. Defaults to false so the feature is by default off and needs
to be explictly set on.

Also skip IPv6 disabling if the transport is a VPN with IPv6. This is a
future-proof case when we start to support multiple VPNs at the same
time.