vpn: Add PreventIPv6DataLeak boolean to D-Bus properties
[vpn] Add PreventIPv6DataLeak boolean to D-Bus properties. JB#53542

Add PreventIpv6DataLeak as a run-time changeable property emitted by
vpn-provider.c. This changes the behavior on whether to disable/block
the feature to disable IP family that is not used (currently only for
disabling IPv6) on the particular provider if the provider does not
have either of the IP families enabled.