vpn-provider: Add PreventIpv6DataLeak to provider bool settings
[vpn-provider] Add PreventIPv6DataLeak to provider bool settings. JB#53542

Support PreventIPv6DataLeak value changing on provider. Emit
PropertyChanged signal to connmand when value is set. This value depends
on the actual VPN setting and does not need to be saved along provider
settings.

Add a helper function for VPN plugins to use for changing the value as
each plugin uses it to simplify code. The function supports enabling of
either or both of the IP family networks but current functionality does
changes only to the PreventIPv6DataLeak value. This is meant to be a
part of future plans.