openvpn: Control IPv6 data leak prevention with --block-ipv6 option
[openvpn] Control IPv6 data leak prevention with --block-ipv6 option. JB#53542

Use the --block-ipv6 option that is added in OpenVPN version >= 2.5
(https://build.openvpn.net/man/openvpn-2.5/openvpn.8.html) to control
the feature to prevent data leak for the VPN if it has IPv6 set as off and
ipconfig disabled. If the option is omitted disable the feature to
prevent data leak and require the option to be explicitly enabled in
order to enable data leak prevention feature.