vpnc: Control IPv6 data leak prevention  with additional option
[vpnc] Control IPv6 data leak prevention with additional option. JB#53542

Add "VPNC.BlockIPv6" option for VPNC to control the IPv6 data leak
prevention feature. If omitted default to false and require the feature
to be explicitly set to true.

VPNC does not have a separate option that is to be used but for
convenience is saved along the VPNC provider configuration.