[openfortivpn] Control IPv6 data leak prevention with additional PPPD noipv6. JB#53542
Add the PPPD supported option noipv6 to be used to control IPv6 data
leak prevention feature. The value must be explicitly set to true to
enable the feature.

OpenFortiVPN does not yet support the option but as it uses PPPD the
option it may be enabled in the future releases.