timerserver: Fix time update manual->auto
__connman_timeserver_conf_update() will block the update if the
service is not matching the current used service.  In case where
ts_service is NULL it should not block the update. This can happen
by a manual to auto TimeUpdate settings change.

While at it reorder the condition in __connman_timeserver_sync() to
match the order in __connman_timeserver_conf_update().

Reported by Emmanuel Vautrin