[vpnc] Fix internal option use to support BlockIPv6 option. JB#54607
Read the "BlockIPv6" option separately out of the loop. VPNC does not
have support for the value and vpnc_opt for this is set as NULL.

Add support for more internal options where vpnc_opt is defined NULL.
Idea is to skip all of these when writing the configuration. This does
not prevent saving as cm_opt is used then.