[openvpn] Ignore all options without value set to false. JB#54606
Ignore the options which have no value and are set to false to comply
with the set value for the option. These options are either set or not
set, so treat them as booleans.

Move "BlockIPv6" use to later to be used directly from provider values.

Require OpenVPN 2.5.2 since "--block-ipv6" is implemented in version
2.5 and above. The option is used only at run-time.