[service] Fix autoconnect back-off timer for cellular networks. JB#52947
The timer was made to trigger autoconnect round for services that are
in failed state. However, the cellular service does not stay in the failed
state for long and goes idle right after the timer is started due to
gofono's "activate-failed" signal being triggered.

This commit makes autoconnect round to start regardless the service is in idle
state. It is safe to do even if the user cancelled the auto-connect.