l2tp: Improve invalid auth and disconnect notify, fix cb use
Stop the task running xl2tpd when it notifies being disconnected as in
case the authentication fails the process keeps on looping with the
notify and cannot be given new credentials until stopped with a D-Bus
call to disconnect. Implement a connect done function to ensure that the
callback is called only once as it sends a reply back to the caller. Use
this function when stopping, having invalid credentials or other errors.

Also, call the callback via connect done only after the connection is
established. This fixes the issue of l2tp reporting back 0 ("ok") before
it gets an response from the connection negotiation, which informed the
caller with completely invalid information.