vpn-provider: Ignore error adding when state is idle/unknown
[vpn-provider] Ignore error adding when state is idle/unknown. JB#54276

Do not allow to add errors for provider that is already set into idle
state or is in unknown state. This case may happen when networks are
rapidly changed and VPN did not call the callback connect_cb() until the
VPN is died and vpn.c:vpn_died() initiates cleanup in the VPN which
eventually calls connect_cb() with an error.