[vpn-provider] Cleanup to be in sync with upstream. Fixes JB#50320
Coding style fixes and refactoring to be more in-line with upstream
code.

Remove the checks for VPN agent to determine whether a VPN can be
connected. Each VPN plugin already does this check prior to attempting
to launch a query for credentials to an agent.

Remove the delayed connect function if connecting a VPN when connmand is
online. It is better to connect asap than to wait the delayed connect to
be scheduled.