[vpn] Pass ENOLINK as a reply to connect request. JB#50320
It is important to pass ENOLINK error that is reported back by vpnd
when connmand is offline back to the caller as a reply to the connect
request. This is because the cb_data->callback() would otherwise left
unhandled in such case as the state is not going to be changed on the
VPN and thus, set_provider_state() will not get triggered. ENOLINK from
vpnd means that connection is not delayed to be connected later but is
refused to be connected. That must be informed to the caller as
"NoCarrier" D-Bus message which this change enables.

Also refactor connect_reply() to clearly show all errors received from
vpnd in readable manner.