[network] Add disconnecting status bool and getter for it. JB#53062
Add a bool for disconnecting status that is set "true" only prior to
connecting the network driver to disconnect. Setting the value to
"false" is done by set_disconnected() which gets eventually called
either by directly via __connman_network_disconnect() or the network
driver calls the connman_network_set_connected() to update status. Also
set the value to "false" when requested to connect.

This is useful for plugins to check in what state the network is.

Added also some more debug calls.