[ofono] Fix ipconfig method use and don't configure disconnecting. JB#53062
Plugin must not change the ipconfig method to DHCP when the original
method is set to be fixed by ofonod. When the modem is connected it
usually does not have all addressing info available yet and the
connecting status is added to modem_configure_ipv4() for a further
check. Otherwise as a result the DHCP is started for the IPv4 and DNS
servers apparently are not being setup, and with consequtive calls from
modem_configure() the method change would require network re-connect as
that is not applied.

Also, utilize connman_network_get_disconnecting() when modem_configure()
is called to not to update the modem addressing and thus, change the
ipconfig method when the IPv4 and/or IPv6 entries are most likely to be
missing, resulting in wrong methods being set up for the cellular.
Without this the settings file would contain inconsistent information
(DHCP being set when disconnecting even though FIXED was previously set
when connected).