[connman] Set systemd service type to dbus, update from upstream. JB#55244
Set the service type of connmand to dbus and apply the service name.
This does seem to reduce the initialization time of connmand from ~700ms
to ~70ms according to systemd-analyze. This varies a bit naturally on
different devices. We've been using this as nofity type since 2014, see
commit e59306cc61ec1807cbd7fc1a2838ecd6764bf120.

Use also same "Conflict" rule to shutdown.target as it is used in
upstream. Set the "Before" to follow upstream service file as well.

Include the "DefaultDependencies" as well to avoid having dependency
cycles in any case in the future.