[network] Add function for checking if network is configured. JB#53062
This function can be used by the plugins to determine whether the network
in use is either configured with a method that auto-assigns local address
or if the local address has been assigned for a FIXED/MANUAL ifconfig.
Uses __connman_ifconfig_is_configured() to check the required ipconfigs
based on the given type.