[ofono] Implement delayed modem connect with bailout timer. JB#53062
Implement a function to delay the connection of the modem if all
required ipconfigs are not configured yet. This is done due to the issue
that when setting cellular connected before the IPv4 address is received
from libofono IPv4 is setup wrong. In such case two IP addresses will be
used if there was a previous one existing in the service settings file,
which in that case is loaded first and then the libofono reports a new
address resulting two different addresses for an interface configured to
use IPv4.

By delaying the connection until the information is received from
libofono the network connection is established with the latest
information reported by the mobile network operator. This applies only
to cases where the method is set either to FIXED or MANUAL

The bailout timer of approx. 2 seconds is added in order to not to make
cellular completely inoperable if the modem or operator or ofono does
not report proper data back. The device can operate in IPv6 only mode,
for instance, but that is not desired and thus, an error is logged on
that occasion.