[ipconfig] Do __connman_ipconfig_clear_address() for FIXED/MANUAL only. JB#53062
Change the __connman_ipconfig_clear_address() to clear the ipaddress
only for ipconfig that is set to be FIXED or MANUAL. Only with these the
existing value may cause confict or invalid use in special cases such as
with cellular that gets the addresses via plugin from an external
component.

This function was not used by any internal source.