[network] Add function to clear the ipaddresses of given type. JB#53062
Allow to clear the ipaddress stored in the ipconfig with the given
ipconfig type. This is useful for plugins that rely on external
component setting the ipaddress information in cases where existing
addresses are most likely invalid and can cause conflict or error when
applied from cache. By using this at the time of setting the network
disconnected the cache is cleared by utilizing
__connman_ipconfig_clear_address().