[ofono] Clear ipaddress when network is disconnected. JB#53062
Clear the ipaddress from cache in ipconfig(s) when setting the network
disconnected. This ensures that the cached value (for the STATIC ->
ipconfig FIXED type) is not used when the network is reconnected. Ofono
must inform this address before it is to be applied. Using the old
address in some cases results in having 2 addresses for the IPv4 when
networks change.

A bit similar issue was happening at bootup but then the cache was empty
and the old value was read from settings file. This ensures the cache is
as well cleaned up.