[connman] Add StorageRoot perms and default to 755. Fixes JB#54987
Add a StorageRootPermissions configuration variable to be used to define
the permissions for the main StorageRoot dir. The StorageRoot is now
created with the set permissions, which defaults to 755. For other
content below the StorageRoot the StorageDirPermissions are used.

This directory is attempted to be created at startup.
__connman_storage_create_dir() will return error only when the dir
cannot be created, no error is returned with EEXIST.