[packaging] Move target.wants symlinks from network to multi-user. JB#55244
It is wrong to set network.target to wait for connmand to start. It
should be the other way around according to systemd documentation
https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/ and
upstream ConnMan uses the service file rules to do so. These are updated
in previous commit. Instead, to follow the "Install" rules create
symlinks for multi-user.target.wants.