[connman] Remove unnecessary --systemd as startup option. JB#55244
The option "--systemd" was from our fork changes from long time ago when
it was decided the systemd service type "notify" was to be used. This
was anyways wrongly named option ("--systemd-notify" would have been
better) and the change is reverted so this option is unnecessary, hence
dbus type is in use in upstream and in major distros (e.g., Debian).