[storage] Prevent symlinks and references in paths. Fixes JB#52090
Prevent using a path that contains symlinks or references by checking
that the path expanded with realpath() matches to the original. This way
we can be sure that the location defined in settings is not actually
pointing to any other directory, e.g., via a symlink to another users'
files.

When the paths match or the path does not exist yet storage.c functions
as normal. In case the path is unusable we continue working but cannot
save anything to the location. This applies to any level in the path, so
if a user attempts to symlink a single service dir or a file into the
users' save path even using root user the use of such content is
prevented.

In order to faciliate this also with saving the umask() calls were set
to be done only when file access is really needed.