technology: Create settings file in user change post command
[technology] Create settings file in user change post command. JB#55593

Move the global offline mode initialization from
__connman_technology_init() into a separate function that can be called
also when the __connman_technology_load_from_config() is called as a
post-user change function by storage.c and there is no settings file for
a new user. This ensures the same functionality as when multi-user is
disabled, the initialization creates default settings file with offline
mode only. This change affects only to new users and the function as a
post() function still returns 'false' as there is nothing to process in
this case, all technologies are still off as changing away from previous
user has disabled them.

With return value false the function still indicates that the post()
call was not complete. In this case it is a bit misleading to printout
the message as warning so changed to debug only, since this is a quite
normal action to do with new users.